homework #3
 changes:
  - change requests: get weather by lat & lon instead of by city id
  - add google places suggest fragment in settings
  - add saving last chosen city in SharedPreferences and put it back
  - double fields in response instead of int because sometimes they are double

+++ app/src/main/java/com/example/kimichael/yamblz_forecast/data/network/forecast/response/Temp.java
@@ -9,7 +9,7 @@ public class Temp {
     @SerializedName("temp")
     private Double temp;
     @SerializedName("pressure")
-    private Integer pressure;
+    private Double pressure;
     @SerializedName("humidity")
     private Integer humidity;
     @SerializedName("temp_min")
@@ -25,11 +25,11 @@ public class Temp {
         this.temp = temp;
     }
 
-    public Integer getPressure() {
+    public Double getPressure() {
         return pressure;
     }
 
-    public void setPressure(Integer pressure) {
+    public void setPressure(Double pressure) {
         this.pressure = pressure;
     }
